requirements.txt
praw==7.7.1
openai>=1.99.0
python-dotenv==1.0.1
pandas==2.2.3
tqdm==4.66.5
click==8.1.7